pyproject.toml
[project]
name = "ehr-writeback"
version = "0.1.0"
description = "Open-source EHR write-back integration framework — closes the loop from analytics to clinical workflow via FHIR"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "SiC Projects" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.0",
    "cryptography>=42.0",
    "pyjwt>=2.8",
]

[project.optional-dependencies]
databricks = [
    "databricks-sdk>=0.20",
    "pyspark>=3.5",
    "delta-spark>=3.1",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "respx>=0.21",
    "ruff>=0.3",
    "mypy>=1.9",
]

[project.urls]
Repository = "https://github.com/slusset/databricks-ehr-integration-core"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ehr_writeback"]

[tool.ruff]
src = ["src"]
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
strict = true
